update deploy_testing and testutils to use *plotbot.Message
I was seeing this really really odd behaviour where reassigning to a
variable with the output of a function would somehow update a
reference in the struct of the variable it was replacing. Changing the
variable *name* fixed the issue. Moved to a pattern where we
explicitly are not reusing variables and are not creating variables we
do no not need to use later on... Also found a testing bug during the refactor