Remove "with" syntax for deploy tags
Also adds `?:` to regexp groups that we don't want to capture.  Capturing
is the default, but when you capture unnecessarily it means every
time you add or remove a group you need to update capture group numbers.
So adding `?:` means that capture group numbers will only change when we
actually change something that needs to be captured, and group numbers
also make more sense.